Get rid of the spurious validate() method.
Currently it's a no-op. Even if we were to implement it,
there's not much it can do other than check that required fields are
present. But 'required' is a silly feature anyway. Plus, if we start
enforcing it now who knows what might break.

Getting rid of this will make implementing unknown field preservation
a bit easier.